chore(core): backend restructuration: config, cms, dialog (#4678)
* chore(dev): restructure backend

* change dialog sessions path

* pr comments

* chore(core): backend restructuration: telemetry (#4679)
