+++ src/bp/core/api.ts
@@ -1,7 +1,14 @@
 import * as sdk from 'botpress/sdk'
+import { CMSService, renderRecursive } from 'core/cms'
+import { ConfigProvider } from 'core/config'
+import Database from 'core/database'
+import { StateManager } from 'core/dialog'
+import { SessionIdFactory } from 'core/dialog/sessions'
 import { EventEngine, EventRepository } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
+import { LoggerProvider } from 'core/logger'
 import { WellKnownFlags } from 'core/sdk/enums'
+import { TYPES } from 'core/types'
 import { inject, injectable } from 'inversify'
 import Knex from 'knex'
 import _ from 'lodash'
@@ -12,30 +19,23 @@ import { isTrainingAlreadyStarted, isTrainingCanceled } from 'nlu-core/errors'
 import modelIdService from 'nlu-core/model-id-service'
 
 import { container } from './app.inversify'
-import { ConfigProvider } from './config/config-loader'
-import Database from './database'
-import { LoggerProvider } from './logger'
+
 import { getMessageSignature } from './misc/security'
-import { renderRecursive } from './misc/templating'
 import { ModuleLoader } from './module-loader'
 import { UserRepository } from './repositories'
 import { Event, RealTimePayload } from './sdk/impl'
 import HTTPServer from './server'
 import { GhostService } from './services'
 import { BotService } from './services/bot-service'
-import { CMSService } from './services/cms'
 import { DialogEngine } from './services/dialog/dialog-engine'
-import { SessionIdFactory } from './services/dialog/session/id-factory'
 import { HookService } from './services/hook/hook-service'
 import { JobService } from './services/job-service'
 import { MediaServiceProvider } from './services/media'
 import { ConversationService } from './services/messaging/conversations'
 import { MessageService } from './services/messaging/messages'
-import { StateManager } from './services/middleware/state-manager'
 import { NotificationsService } from './services/notification/service'
 import RealtimeService from './services/realtime'
 import { WorkspaceService } from './services/workspace-service'
-import { TYPES } from './types'
 
 const http = (httpServer: HTTPServer) => (identity: string): typeof sdk.http => {
   return {
