+++ src/bp/core/app.inversify.ts
@@ -1,5 +1,6 @@
 import { Logger } from 'botpress/sdk'
 import { EventCollector } from 'core/events'
+import { TelemetryContainerModules } from 'core/telemetry'
 import { Container } from 'inversify'
 
 import { BotpressAPIProvider } from './api'
@@ -16,7 +17,6 @@ import { MigrationService } from './services/migration'
 import { DataRetentionJanitor } from './services/retention/janitor'
 import { DataRetentionService } from './services/retention/service'
 import { ServicesContainerModules } from './services/services.inversify'
-import { TelemetryContainerModules } from './services/telemetry/telemetry.inversify'
 import { WorkspaceService } from './services/workspace-service'
 import { Statistics } from './stats'
 import { TYPES } from './types'
