+++ src/bp/core/config/index.ts
@@ -1,0 +1,4 @@
+export * from './bot.config'
+export * from './botpress.config'
+export * from './config-loader'
+export * from './config-utils'
