+++ src/bp/core/database/tables/index.ts
@@ -1,14 +1,15 @@
 import { WorkspaceInviteCodesTable, WorkspaceUsersTable } from 'core/collaborators'
+import { DialogSessionTable } from 'core/dialog/sessions/dialog_sessions-table'
 import { EventsTable } from 'core/events/event-table'
 import { KeyValueStoreTable } from 'core/kvs/kvs-table'
 import { LogsTable } from 'core/logger/logs-table'
+import { TelemetryTable } from 'core/telemetry/telemetry-table'
 import Knex from 'knex'
 
 import { Table } from '../interfaces'
 
 import {
   BotUsersTable,
-  DialogSessionTable,
   ConversationsTable,
   MessagesTable,
   GhostFilesTable,
@@ -16,13 +17,7 @@ import {
   NotificationsTable,
   TasksTable
 } from './bot-specific'
-import {
-  ChannelUsersTable,
-  DataRetentionTable,
-  MigrationsTable,
-  ServerMetadataTable,
-  TelemetryTable
-} from './server-wide'
+import { ChannelUsersTable, DataRetentionTable, MigrationsTable, ServerMetadataTable } from './server-wide'
 
 const tables: typeof Table[] = [
   ServerMetadataTable,
