+++ src/bp/core/repositories/index.ts
@@ -1,3 +1,2 @@
 export * from './notifications'
-export * from './sessions'
 export * from './users'
