+++ src/bp/core/server.ts
@@ -6,8 +6,11 @@ import { RequestWithUser } from 'common/typings'
 import compression from 'compression'
 import cookieParser from 'cookie-parser'
 import session from 'cookie-session'
+import { CMSService, CMSRouter } from 'core/cms'
+import { ExternalAuthConfig, ConfigProvider } from 'core/config'
 import { ConverseRouter, ConverseService } from 'core/converse'
 import { LogsService, LogsRepository } from 'core/logger'
+import { TelemetryRouter, TelemetryRepository } from 'core/telemetry'
 import cors from 'cors'
 import errorHandler from 'errorhandler'
 import { UnlicensedError } from 'errors'
@@ -29,20 +32,14 @@ import portFinder from 'portfinder'
 import { URL } from 'url'
 import yn from 'yn'
 
-import { ExternalAuthConfig } from './config/botpress.config'
-import { ConfigProvider } from './config/config-loader'
-
 import { ModuleLoader } from './module-loader'
-import { TelemetryRepository } from './repositories/telemetry'
 import { AdminRouter, AuthRouter, BotsRouter, MediaRouter, ModulesRouter } from './routers'
-import { ContentRouter } from './routers/bots/content'
 import { HintsRouter } from './routers/bots/hints'
 import { NLURouter } from './routers/bots/nlu'
 import { isDisabled } from './routers/conditionalMiddleware'
 import { InvalidExternalToken, PaymentRequiredError } from './routers/errors'
 import { SdkApiRouter } from './routers/sdk/router'
 import { ShortLinksRouter } from './routers/shortlinks'
-import { TelemetryRouter } from './routers/telemetry'
 import { hasPermissions, monitoringMiddleware, needPermissions } from './routers/util'
 import { GhostService } from './services'
 import ActionServersService from './services/action/action-servers-service'
@@ -52,7 +49,6 @@ import { AuthStrategies } from './services/auth-strategies'
 import AuthService, { EXTERNAL_AUTH_HEADER, SERVER_USER, TOKEN_AUDIENCE } from './services/auth/auth-service'
 import { generateUserToken } from './services/auth/util'
 import { BotService } from './services/bot-service'
-import { CMSService } from './services/cms'
 
 import { FlowService } from './services/dialog/flow/service'
 import { SkillService } from './services/dialog/skill/service'
@@ -90,7 +86,7 @@ export default class HTTPServer {
   private readonly authRouter: AuthRouter
   private readonly adminRouter: AdminRouter
   private readonly botsRouter: BotsRouter
-  private contentRouter!: ContentRouter
+  private cmsRouter!: CMSRouter
   private nluRouter!: NLURouter
   private readonly modulesRouter: ModulesRouter
   private readonly shortLinksRouter: ShortLinksRouter
@@ -245,7 +241,7 @@ export default class HTTPServer {
 
     await this.mediaRouter.initialize()
     await this.botsRouter.initialize()
-    this.contentRouter = new ContentRouter(
+    this.cmsRouter = new CMSRouter(
       this.logger,
       this.authService,
       this.cmsService,
@@ -255,7 +251,7 @@ export default class HTTPServer {
     this.nluRouter = new NLURouter(this.logger, this.authService, this.workspaceService, this.nluService)
     this.converseRouter = new ConverseRouter(this.logger, this.converseService, this.authService, this)
     this.hintsRouter = new HintsRouter(this.logger, this.hintsService, this.authService, this.workspaceService)
-    this.botsRouter.router.use('/content', this.contentRouter.router)
+    this.botsRouter.router.use('/content', this.cmsRouter.router)
     this.botsRouter.router.use('/converse', this.converseRouter.router)
     this.botsRouter.router.use('/nlu', this.nluRouter.router)
