+++ src/bp/core/telemetry/stats/actions-stats.ts
@@ -4,17 +4,16 @@ import LicensingService from 'common/licensing-service'
 import { buildSchema, TelemetryEvent } from 'common/telemetry'
 import Database from 'core/database'
 import { calculateHash } from 'core/misc/utils'
-import { TelemetryRepository } from 'core/repositories/telemetry'
+import { GhostService } from 'core/services'
+import { BotService } from 'core/services/bot-service'
+import { FlowService } from 'core/services/dialog/flow/service'
+import { JobService } from 'core/services/job-service'
 import { TYPES } from 'core/types'
 import { inject, injectable } from 'inversify'
 import _ from 'lodash'
 import ms from 'ms'
 
-import { GhostService } from '..'
-import { BotService } from '../bot-service'
-import { FlowService } from '../dialog/flow/service'
-import { JobService } from '../job-service'
-
+import { TelemetryRepository } from '../telemetry-repository'
 import { TelemetryStats } from './telemetry-stats'
 
 interface RawFlow {
