+++ src/bp/core/telemetry/stats/configs-stats.ts
@@ -2,23 +2,21 @@ import { BotDetails } from 'botpress/sdk'
 import { BUILTIN_MODULES } from 'common/defaults'
 import LicensingService from 'common/licensing-service'
 import { buildSchema, TelemetryEvent } from 'common/telemetry'
-import { BotConfig } from 'core/config/bot.config'
-import { BotpressConfig } from 'core/config/botpress.config'
-import { ConfigProvider } from 'core/config/config-loader'
+import { BotConfig, BotpressConfig, ConfigProvider } from 'core/config'
 import Database from 'core/database'
 import { calculateHash } from 'core/misc/utils'
 import { ModuleLoader } from 'core/module-loader'
-import { TelemetryRepository } from 'core/repositories/telemetry'
+import { GhostService } from 'core/services'
+import { BotService } from 'core/services/bot-service'
+import { JobService } from 'core/services/job-service'
+import { Config } from 'core/services/module/config-reader'
 import { TYPES } from 'core/types'
 import { inject, injectable } from 'inversify'
 import defaultJsonBuilder from 'json-schema-defaults'
 import _ from 'lodash'
 import ms from 'ms'
 
-import { GhostService } from '..'
-import { BotService } from '../bot-service'
-import { JobService } from '../job-service'
-import { Config } from '../module/config-reader'
+import { TelemetryRepository } from '../telemetry-repository'
 
 import { DEFAULT, REDACTED, TelemetryStats } from './telemetry-stats'
