+++ src/bp/core/telemetry/stats/legacy-stats.ts
@@ -1,9 +1,14 @@
 import LicensingService from 'common/licensing-service'
 import { machineUUID } from 'common/stats'
-import { ConfigProvider } from 'core/config/config-loader'
+import { CMSService } from 'core/cms'
+import { ConfigProvider } from 'core/config'
 import Database from 'core/database'
 import { UserRepository } from 'core/repositories'
-import { TelemetryRepository } from 'core/repositories/telemetry'
+import { GhostService } from 'core/services'
+import AuthService from 'core/services/auth/auth-service'
+import { BotService } from 'core/services/bot-service'
+import { JobService } from 'core/services/job-service'
+import { WorkspaceService } from 'core/services/workspace-service'
 import { TYPES } from 'core/types'
 import { inject, injectable } from 'inversify'
 import ms from 'ms'
@@ -11,13 +16,7 @@ import os from 'os'
 import uuid from 'uuid'
 import yn from 'yn'
 
-import { GhostService } from '..'
-import AuthService from '../auth/auth-service'
-import { BotService } from '../bot-service'
-import { CMSService } from '../cms'
-import { JobService } from '../job-service'
-import { WorkspaceService } from '../workspace-service'
-
+import { TelemetryRepository } from '../telemetry-repository'
 import { TelemetryStats } from './telemetry-stats'
 
 @injectable()
